# docs/api-specs.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Eduflex RESTful API Documentation 1.0",
    "description": "Ini adalah dokumentasi RESTful API untuk kursus online Eduflex.",
    "version": "1.0.1"
  },
...
  "servers": [
    {
      "url": "https://api-eduflex.cloud"
    }
  ],
...
  "paths": {
...
    "/users": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "Get current user in active session",
        "description": "",
        "operationId": "getUser",
        "parameters": [
          {
              "name": "Authorization",
              "in": "header",
              "required": true,
              "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer",
                      "example": 1
                    },
                    "name": {
                      "type": "string"
                    },
                    "role": {
                      "type": "string"
                    },
                    "profession": {
                      "type": "string"
                    },
                    "avatar": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Invalid token"
          }
        }
      },
      "put": {
        "tags": [
          "user"
        ],
        "summary": "Update user",
        "description": "This can only be done by the logged in user.",
        "operationId": "updateUser",
        "parameters": [
          {
              "name": "Authorization",
              "in": "header",
              "required": true,
              "type": "string"
          }
        ],
        "requestBody": {
          "description": "Update an existent user in the store",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  },
                  "profession": {
                    "type": "string"
                  },
                  "avatar": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success update user",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "number"
                        },
                        "name": {
                          "type": "string"
                        },
                        "email": {
                          "type": "string"
                        }
                      }
                    },
                    "errors": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "JWT must be provided"
          }
        }
      }
    },
...
    "/media": {
      "get": {
        "tags": [
          "media"
        ],
        "summary": "Show all uploaded media",
        "description": "",
        "operationId": "getUserByName",
        "parameters": [
          {
              "name": "Authorization",
              "in": "header",
              "required": true,
              "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer"
                    },
                    "image": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "JWT must be provided"
          }
        }
      },
      "post": {
        "tags": [
          "media"
        ],
        "summary": "Upload media",
        "description": "Upload media in base64 format",
        "operationId": "uploadMedia",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "image": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer"
                    },
                    "image": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "JWT must be provided"
          }
        }
      }
    },
    "/media/{id}": {
      "delete": {
        "tags": [
          "media"
        ],
        "summary": "Delete media",
        "description": "Delete media from database",
        "operationId": "deleteMedia",
        "parameters": [
          {
              "name": "Authorization",
              "in": "header",
              "required": true,
              "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Image has been deleted!"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "JWT must be provided"
          }
        }
      }
    },
...
    "/courses/{id}": {
      "put": {
        "tags": [
          "courses"
        ],
        "summary": "Update courses",
        "description": "This operation can be done only for admin.",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string"
          }
        ],
        "requestBody": {
          "description": "These attributes are optional. You can update 1/2 attribute.",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "certificate": {
                    "type": "boolean"
                  },
                  "thumbnail": {
                    "type": "string"
                  },
                  "type": {
                    "type": "string"
                  },
                  "status": {
                    "type": "string"
                  },
                  "price": {
                    "type": "integer"
                  },
                  "level": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "mentor_id": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Course updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer"
                        },
                        "mentor_id": {
                          "type": "integer"
                        },
                        "name": {
                          "type": "string"
                        },
                        "certificate": {
                          "type": "boolean"
                        },
                        "thumbnail": {
                          "type": "string"
                        },
                        "type": {
                          "type": "string"
                        },
                        "status": {
                          "type": "string"
                        },
                        "price": {
                          "type": "integer"
                        },
                        "level": {
                          "type": "string"
                        },
                        "description": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "JWT must be provided"
          }
        }
      },
      "delete": {
        "tags": [
          "courses"
        ],
        "summary": "Delete courses",
        "description": "This operation can be done only for admin.",
        "parameters": [
          {
              "name": "Authorization",
              "in": "header",
              "required": true,
              "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Course has been deleted!"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "JWT must be provided"
          }
        }
      }
    },
...
    "/chapters/{id}": {
      "get": {
        "tags": [
          "chapters"
        ],
        "summary": "Show courses by id",
        "description": "This operation can be done only for admin.",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Chapter updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer"
                        },
                        "name": {
                          "type": "string"
                        },
                        "course_id": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "JWT must be provided"
          }
        }
      },
      "put": {
        "tags": [
          "chapters"
        ],
        "summary": "Update chapter",
        "description": "This operation can be done only for admin.",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string"
          }
        ],
        "requestBody": {
          "description": "These attributes are optional. You can update 1/2 attribute.",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Chapter updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer"
                        },
                        "name": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "JWT must be provided"
          }
        }
      },
      "delete": {
        "tags": [
          "chapters"
        ],
        "summary": "Delete chapter",
        "description": "This operation can be done only for admin.",
        "parameters": [
          {
              "name": "Authorization",
              "in": "header",
              "required": true,
              "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Course has been deleted!"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "JWT must be provided"
          }
        }
      }
    },
...
  }
}